+++ src/test/java/org/folio/rest/impl/StorageTestSuite.java
@@ -35,6 +35,7 @@ import io.vertx.core.logging.LoggerFactory;
 
 @Suite.SuiteClasses({
   CrudTest.class,
+  InvoiceTest.class,
   InvoiceNumberTest.class,
   InvoiceLineNumberTest.class,
   VoucherNumberTest.class
